fix: use global packages for NixOS attached homes
Previously we didn't use global packages, which meant that if you had
options defined in your nilla inputs for nixpkgs (to allow unfree
packages, say) these would not be respected.

To fix this, we can use home-manager.useGlobalPkgs. Setting this option
as a workaround was suggested to me by SqueakyBeaver - thanks!

Co-Authored-By: Beaverr <squeaky.beaver4133@protonmail.com>